Fix: cannot make offers to direct listings with native token
